# Using Interlocked in InputFlowControl and Http2Stream